neonvm: clock synchronization using kvm_ptp (#732)
Synchronizing VM clocks to host using kvm_ptp. We enable PTP clock (and
the KVM related directive) on the kernel and use chrony on the VM as a
server. More info
[here](https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/7/html/virtualization_deployment_and_administration_guide/chap-kvm_guest_timing_management#sect-KVM_guest_timing_management-Host-guest-time-sync).

`chronyd` gradually synchronizes the clock by speeding it up or down. If
the clock drift is large then this might take a long time. However, here
we avoid jumps in the time, which is not ideal from user's perspective.
Chrony's progress can be tracked using `chronyc tracking` command.

Fixes https://github.com/neondatabase/autoscaling/issues/712