agent: Send compute unit in requests to plugin
Part of moving compute unit source of truth from the scheduler plugin to
the autoscaler-agent.

We also particularly want to have this so that the plugin doesn't
approve resources that aren't an integer multiple of the respective
value in the compute unit. If we didn't we could e.g. approve 1.01 CPUs,
leaving the autoscaler-agent with 0.01 CPUs that it'll never use.
In practice, that's quite unlikely to happen, but it's still worth
avoiding potential issues there.