neonvm: Use container statuses, not pod phase, to trigger restart
Extracted from #738, which adds a second container to the runner pods.
Because of that second container, if only one container exits, the pod
will still have .status.phase = Running, so we need to proactively
notice that one of the containers has stopped and propagate that status
to the VM itself.

This also introduces some funky logic around how we handle restarts:
Because the 'Succeeded' and 'Failed' phases no longer imply that QMEU
itself has stopped, we need to explicitly wait until either the pod is
gone or the neonvm-runner container has stopped; otherwise we could end
up with >1 instance of the VM running at a time.