agent,plugin: Remove unused metrics from api.Metrics
Ref https://github.com/neondatabase/autoscaling/pull/737#discussion_r1458396094

Basically, with the introduction of a secondary metrics type for
internal use inside the autoscaler-agent, we can get rid of the other
metrics that aren't used by the scheduler plugin.

This PR is the first step (making the fields pointers and omitempty);
we'll have to do a follow-up to actually remove the fields.