neonvm-controller: Fix overwriting runner version
Noticed while working on #738. In short, because the runner API version
was part of labelsForVirtualMachine, any update to the runner version
would be updated for *all* VM pods, not just new one.

This is (probably) not an issue in prod right now, but could be an issue
for future changes. Right now, #738 would trigger this bug, but remain
unaffected (it doesn't have version-dependent logic *currently*).

This PR fixes the behavior by adding the runner API version as an
explicit argument to labelsForVirtualMachine and ignoring the label in
updatePodMetadataIfNecessary.