neonvm-controller: Fix overwriting runner version (#753)
Noticed while working on #738. In short, because the runner API version
was part of labelsForVirtualMachine, any update to the runner version
would be updated for *all* VM pods, not just new one.

This is (probably) not an issue in prod right now, but could be an issue
for future changes.

This PR fixes the behavior by adding the runner API version as an
explicit argument to labelsForVirtualMachine and ignoring the label in
updatePodMetadataIfNecessary.