plugin: Handle panics from agent requests (#785)
Closes #760.

AFAIK this hasn't been an issue in the past, but as we're trying to
improve reliability, it's good to get this out of the way before it
becomes an issue.

Note that this PR is quite minimal - expanding the existing tech debt we
have around how the scheduler plugin handles HTTP requests.
It's probably ok *enough* for now. I don't expect we'll be making too
many changes to it in the near future. See also: #13.

Tested locally by forcing it to panic on every request:

diff --git a/pkg/plugin/run.go b/pkg/plugin/run.go
index 007554a..6da7728 100644
--- a/pkg/plugin/run.go
+++ b/pkg/plugin/run.go
@@ -262,8 +262,10 @@ func (e *AutoscaleEnforcer) handleAgentRequest(
 		}
 	}

-	pod.vm.mostRecentComputeUnit = &e.state.conf.ComputeUnit
-	return &resp, 200, nil
+	panic(errors.New("test panic!"))
+
+	// pod.vm.mostRecentComputeUnit = &e.state.conf.ComputeUnit
+	// return &resp, 200, nil
 }

 // getComputeUnitForResponse tries to return compute unit that the agent supports

The change appears to work as intended.