neonvm: Merge kustomize patches into source yamls (#824)
Now that we don't actually need to have different configurations for
"default" vs "default-vxlan", we can just put the content of the config
patches back into the yamls defining the objects they modify, keeping
things simpler.

We had to keep webhookcainjection_patch because the webhook manifests
are generated via `make manifests`, so we have to modify it after.

Only rendered yaml diff is swapping the ordering of the
neonvm-controller container's port definitions:

3803,3805d3802
<         - containerPort: 9443
<           name: webhook-server
<           protocol: TCP
3807a3805,3807
>           protocol: TCP
>         - containerPort: 9443
>           name: webhook-server