plugin: Use RelistIfHaventSince(..) in getVmInfo
This change passes the VM store to the pod watcher (which requires
reordering them, and initializing the VM watcher's copy of the pod store
later), as well as passing the watch.RelistCount through all of the
functions to get to getVmInfo.

The main reason for this change is that it completely stops the positive
feedback loop it's possible to hit when stale VM pod start events are
handled after the VM was deleted, leading to delays caused by Relist(),
leading to further stale events.

With this change, the amount of required Relist()s are determined not by
the number of events, but by the actual time it takes to complete the
Relist() (a new one is only necessary if an event came in after the
previous one completed - so events are delayed by at most 2x the
duration of a single Relist() call, assuming negligible other handling).