plugin: Use multiple event handler threads
Each handler has its own queue, and we distribute events by hashing the
name of the object we're sending through, so that pod start and stop
event ordering is preserved by going in through the same queue.

This will help reduce the impact of any future issues where the event
queue gets backed up.

Part of #851.