plugin: Remove vmStore from AutoscaleEnforcer (#904)
After the changes from #865 that switched to fetching VmInfo from Pod
objects, the VM store is no longer needed long-term, just during initial
cluster state reading (which can *also* be swapped out) and for event
generation.

This commit changes the VM store from being stored in the
AutoscaleEnforcer struct to being passed in to readClusterState as an
argument.