Generate manifests via `make generate`, remove `make manifests`
We currently see issues with Go >1.21 and our current controller-gen
version, so a reasonable workaround is to always run with Go 1.21
instead of using whatever the system happens to have install.

Easiest way to do that is by doing the manifest generation inside a
container, and `make generate` already serves that purpose, so this
commit changes the generation script to *also* generate the manifests.

This commit also removes all prior occurrences of `make manifests`.

Fixes #885.