plugin: Add transactions for speculative operations (#936)
This commit adds 'pkg/util/xact' with usage in scheduler plugin such
that all places that previously implemented checks for "can this pod fit
on that node" are now done by speculatively reserving the pod, checking
if it left the node over-full, and then deciding whether to accept it.

This is unified through a new method speculativeReserve(). Internally,
speculativeReserve() calls a new method on resourceTransitioner, which
makes follow-up work easier.