neonvm-controller: Rework zap setup & disable sampling
By default, zap enables "log sampling" when not in development mode.
Unsurprisingly, we're finding that neonvm-controller logs are being
dropped in production as a result of this.

I couldn't find a clean way to disable sampling without entirely
constructing the zap.Logger from scratch, so that's what I did.

Notable changes:

* neonvm-controller's '--zap-*' flags have been removed
* neonvm-controller now hard-codes the logger config we were using
  before -- info log level, stack traces on panic, and ISO8601 time