neonvm: Switch to kernel 6.1.92 (#950)
As of 2024-06-03, www.kernel.org says this is the most recent version of
the 6.1.x LTS branch.

Basically just followed the instructions in neonvm/README.md, copying in
linux-config-6.1.63 as .config, and .config back to /host once done with
menuconfig.

Trying this as a partial solution to neondatabase/cloud#10622. It
doesn't *completely* fix it, but it should dramatically improve the
situation.