Support ScaleRequest message, to allow vm-monitor to control autoscaling
The new protocol message allows the vm-monitor to directly specify the
desired size of the VM. With that, the agent doesn't need the metrics
anymore, it will just try to make the vm-monitor's wish true.

This is the autoscaler agent implementation of the RFC I proposed
here: https://github.com/neondatabase/neon/pull/8111. In order to use
the new API, see the corresponding VM monitor changes at:
https://github.com/neondatabase/neon/tree/heikki/wip-autoscale-api