neonvm: Add support for virtio-mem (#970)
Adapted from #705.

This implementation is focused on making virtio-mem a drop-in
replacement for our existing DIMM hotplug -based memory scaling.

As such, the externally visible differences are:

1. VirtualMachines have a new optional field: .spec.guest.memoryProvider
2. VirtualMachines have a new optional field: .status.memoryProvider
3. VirtualMachine .status.memorySize changes in smaller increments if
   using virtio-mem.

This change introduces the concept of a "memory provider", which is
either "DIMMSlots" or "VirtioMem".

If a VM is created without setting the memory provider in the spec, the
controller will use the default one, which is specified by its
--default-memory-provider flag.
This will only persist for the lifetime of the runner pod. If the VM is
restarted, then it may change to a new default.

Additionally, because our usage of virtio-mem has 8MiB block sizes, we
actually *don't* default to virtio-mem if the VM was created with some
odd memory slot size that isn't divisible by 8MiB.

The memory provider of the current runner pod is stored in the new
status field. For old VMs created before this change, the status field
is set to DIMMSlots iff the pod name is set without the status field.

Also, we allow mutating .spec.guest.memoryProvider to allow the
flexibility to solidify a particular provider from future changes or to
force it to change across restarts. There was more discussion on this in
the PR comments, if you're curious.

Co-authored-by: Andrey Taranik <andrey@cicd.team>
Co-authored-by: Oleg Vasilev <oleg@neon.tech>