neonvm: Use memory_hotplug.online_policy=auto-movable for virtio-mem (#981)
In short: Currently we hotplug all memory as ZONE_MOVABLE. This has
known issues because there are theoretical and practical limits on the
ratio between ZONE_MOVABLE and ZONE_NORMAL.
With 'memory_hotplug.online_policy=auto-movable', we can use the config
value 'memory_hotplug.auto_movable_ratio' to set a ratio of MOVABLE:NORMAL
to uphold during hotplug and the kernel will Do The Right Thing™.

This change adds a new flag to neonvm-controller and neonvm-runner to
set the value of this ratio: '-memhp-auto-movable-ratio'.

However, please note: The new behavior is only available for VMs using
virito-mem because memory_hotplug.auto_movable_ratio does not count
hotplugged ZONE_NORMAL DIMM slot towards the ratio, so IMO it's not
worth changing the existing behavior. [1] [2] [3]

This is the productionized version of the experimentation from #959.
For more on that, see [4].

[1]: https://docs.kernel.org/admin-guide/mm/memory-hotplug.html#module-parameters
[2]: https://lwn.net/Articles/865618/
[3]: https://neondb.slack.com/archives/C03TN5G758R/p1718225978299459?thread_ts=1718158200.324759
[4]: https://www.notion.so/neondatabase/59c9e9b2619e4ccd8fbb99e9b0cb1a89