neonvm-controller: Add '--skip-update-validation-for' (#990)
This flag is a new escape hatch to allow arbitrary updates to an object,
bypassing the webhook update validation (but still within the CRD).

Sometimes when trialing things on staging, we end up with objects in a
bad state that can't be deleted because update validation fails when
removing the finalizer (e.g., due to destructive CRD changes).

---

To test, I used vm-deploy.yaml and tried `kubectl edit`ing the maximum
memory. Without this PR, I got:

  error: virtualmachines.vm.neon.tech "postgres16-disk-test" could not be patched: admission webhook "vvirtualmachine.kb.io" denied the request: .spec.guest.memorySlots.max is immutable
  You can run `kubectl replace -f /tmp/kubectl-edit-919695070.yaml` to try this update again.

With this PR, and '--skip-update-validation-for=postgres16-disk-test':

  virtualmachine.vm.neon.tech/postgres16-disk-test edited