neonvm: Fix double prefix in image loader
Basically, we were applying the 'neonvm-' prefix twice when we added
neonvm-runner-image-loader into neonvm.yaml.

The diff in neonvm.yaml from the #988 to now:

  4007c4007
  <   name: neonvm-neonvm-runner-image-loader
  ---
  >   name: neonvm-runner-image-loader

(i.e., the name of the daemonset is corrected, as it should be)

And, there is no diff in neonvm-runner-image-loader.

---

Medium-term, we should probably remove the namespace prefix thing,
but I'd rather just get this fixed ASAP.