Silence "Unknown parameter 'mode'" messages at neonvm startup
The kernel prints a few lines like these to the kernel log at startup:

    [    0.886717] ext3: Unknown parameter 'mode'
    [    0.887279] ext2: Unknown parameter 'mode'
    [    0.887595] ext4: Unknown parameter 'mode'
    [    0.887906] squashfs: Unknown parameter 'mode'

It's because the runtime, ssh-authorized-keys image and any ConfigMaps
or Secrets are iso9660 images, mounted with commands like this:

    /neonvm/bin/mount -o ro,mode=0644 $(/neonvm/bin/blkid -L ssh-authorized-keys) /mnt/ssh

Because we don't specify the filesystem type, the command tries
different filesystems until it finds one that recognizes it. You get
the errors, because "mode" option is not supported by those other
filesystems that the command tries first.

The messages are harmless, but add noise to the logs. To silence them,
specify the "iso9660" filesystem explicitly in the mount command.