agent/core: Implement LFC-aware scaling
Part of #872.
This builds on the metrics that will be exposed by neondatabase/neon#8298.

For now, we only look at the working set size metrics over various
evenly-spaced windows (all 1 minute apart).

The algorithm is somewhat straightforward to implement (see wss.go), but
unfortunately seems to be difficult to understand *why* it's expected to
work.

For more context, refer to the RFC here:
https://www.notion.so/neondatabase/cca38138fadd45eaa753d81b859490c6