Implement scaling latency metrics through revisions (#983)
Newly introduced `Revision` is (basically) an integer value, which can
be associated with different parts of a system: Monitor, Plugin, NeonVM
and scaling algorithm itself. There are two types of such association:
- `TargetRevision` corresponds to the desired state of a particular
part.
- `CurrentRevision` corresponds to the state which was already achieved.

As the system makes progress `TargetRevision` is propagated to
`CurrentRevision` field, simultaneously tracking how long it took for
this propagation.

When the same revision value is passed through multiple parts, we can
measure end-to-end latency of multi-component operations.

Fixes #594.

---------

Signed-off-by: Oleg Vasilev <oleg@neon.tech>