agent: Don't include VM name in neonvm request metric labels
Currently, we sometimes get neonvm request failure metric labels like

  [error: virtualmachines.vm.neon.tech "compute-rapid-wave-a45g5t85" not found]

Thankfully *in practice* the rates are very low, so it's not a big
issue, but in theory we could accidentally blow up our metrics storage
if some issue was causing our requests to fail for every object in a way
that included the object name.

So this PR makes sure we filter out the name.