vm-builder: Run all processes in a cgroup
More specifically, with this change:

1. There is a new cgroup inside the VM, /neonvm-root
2. There are associated cgroup & mount namespaces such that when inside
   those namespaces, /neonvm-root appears to be the root cgroup
3. All 'command's from the image spec are run inside the /neonvm-root
   cgroup and associated namespaces

In order to support the creation of new cgroups inside /neonvm-root
(like what's used in the test images, or in neondatabase/neon), we have
to run normal processes in a separate /neonvm-root/leaf cgroup in order
to not voilate the "no internal processes" rule.[1]

But other than that, most things should continue to behave as normal.
For example, cgconfigparser (which we use to set up user cgroups)
appears to continue to work just fine.

Also worth noting: Currently sshd also runs in the same cgroup &
namespaces as everything else, so that it continues to match the
perspective of processes running inside the VM.

When running as root, it's possible to break out of the namespaces by
entering PID 1's namespaces:

  nsenter --target=1 --cgroup --mount <command...>

[1]: https://man7.org/linux/man-pages/man7/cgroups.7.html