neonvm: Add controller flag to disable runner cgroup
This commit adds new flags to neonvm-controller and neonvm-runner:

* controller: -disable-runner-cgroup
* runner: -enable-dummy-cpu-server

If the controller is passed -disable-runner-cgroup, then it will pass
both -skip-cgroup-management and -enable-dummy-cpu-server to new runner
pods.

If neonvm-runner is passed -enable-dummy-cpu-server (requires the
-skip-cgroup-management flag), it will *not* run QEMU in a cgroup, but
still provide an HTTP server with endpoints that pretend as if the CPU
limit was updated successfully.

Internally, the runner's "dummy" implementation still needs to store the
most recently set value to provide it back to the controller, so that it
doesn't infinitely loop trying to set the CPU.

ref https://neondb.slack.com/archives/C06SJG60FRB/p1723485730034399