plugin/trans: Fix over-allowing from forceApprovalMinimum (#1025)
Noticed as part of #1023.

The current (incorrect) behavior is that forceApprovalMinimum is
relative to the current resources when it's actually intended (and
documented as such) to be absolute. So in practice, we're approving
autoscaler-agent requests that we shouldn't -- allowing resource
increases above the node maximum when we don't need to.

forceApprovalMinimum was added as part of #936, and presumably we didn't
catch this because of the lack of unit tests.