neonvm-controller: Remove support for DIMM slots (#1070)
This is part 1 of 2 for removing DIMM slots support and completing our
transition to using virtio-mem instead (ref #1060).

With this change, neonvm-controller will always assume that VMs are
using virtio-mem. The only checks exist at the top of the reconcile
functions, making sure that this is indeed the case.

The fields still exist in the CRD (although the're marked as
deprecated), so that rollback of this change is still sound.

neonvm-runner also retains support for both memory providers, so that
custom neonvm-runner images continue to work for at least one version
across the change.