Use go-chef for image builds (#989)
It *dramatically* reduces the amount of time it takes to re-run image
building if the set of packages used hasn't changed.
It's hard to get an objective measurement (because we do get *some*
caching from being careful what we add), but:

On my machine, for a rebuild after a change in 'pkg/api', it reduces
`make docker-build` time from 249s -> 22s.

And, it also slightly improves the time it takes to run a clean build.
After `docker system prune -af`, it reduces `make docker-build` on my
machine from 345s -> 126s.