agent,billing: Move clients and queue into new pkg/reporting
This commit adds a new package: pkg/reporting.

The new package contains both the HTTP/S3/Azure Blob client
implementations from pkg/billing, as well as the event queue and sender
implementation from pkg/agent/billing.

The intent is to provide a 'plug and play' infrastructure for
asynchronously sending billing events (among other event types) to a
variety of sources, with the code responsible for generating those
events only being responsible for setting up the clients and queueing
the events.