neonvm-controller: Use virtio-mem when status.memoryProvider is nil
Part of #1060. Before we stop setting status.memoryProvider, we need to
make sure that an empty status.memoryProvider is interpreted as
virtio-mem.

We *could* do that everywhere status.memoryProvider is used, but it's
probably just easier to do it at top, replacing our handling that
originally set DIMM slots by default.