neonvm-controller: Remove unused auth proxy rbac
We used to run kube-auth-proxy as a sidecar alongside neonvm-controller.
That was removed in #825 - I guess the RBAC for it was forgotten.

This showed up because the v0.36.0 release YAMLs have a regression where
the RBAC for this unused proxy refers to 'system' and 'controller'
instead of 'neonvm-system' and 'neonvm-controller'.