neonvm-controller: Have owner refs for migration source and target (#1194)
From now on, during the migration we will have:

- Symmetrical owner refs: both VirtualMachine and
  VirtualMachineMigration will own both source and target pods.

- Asymmetrical controller refs: VirtualMachine controls source pod,
  VirtualMachineMigration controls target pod.

Once the migration is complete, neither the VirtualMachine nor the
VirtualMachineMigration will control the source pod, but they will still
have (non-controlling) owner refs.

This allows us to figure out the pod's role in the migration by
looking at it's references.

Co-authored-by: Oleg Vasilev <oleg@neon.tech>