neonvm-controller: Use non-controller owner ref for migration source
In short, there's currently no reliable way to tell whether a Pod is the
source or target in a live migration without checking against the
migration object itself.

This is because the migration object initially has a unique "controller"
reference on the target pod (with the source pod having a
non-controlling owner reference), and once the migration is complete,
the source pod is changed to have the unique controller reference setup.

So this change is *just* to keep using a non-controller reference on the
source pod after the migration completes.

That means that we will have the following guarantees:

* If a Pod has a "controller" owner reference for a live migration, it's
  the target pod in an ongoing migration.
* If a Pod has a non-controller owner reference for a live migration,
  it's the source pod in a maybe-ongoing migration.
* If the Pod has a "controller" owner reference for a virtual machine,
  it's the current source of the VM. So if a migration source pod
  *doesn't* have an owner reference for the VM, then the migration has
  completed and there is no longer a use for the old source pod.