cluster-autoscaler: Upgrade to k8s 1.29 (#1151)
This patch was essentially a rewrite of the original. 

Context being, upstream, the cluster autoscaler team unified the watch
and lister interfaces using informers. So, instead of patching both of
those methods, only the `List()` method on the `allPodLister` is
modified. Assuming my understanding is correct it'll have the same
behaviour but with much less code (as evidenced by the removed lines in
the original patch)

I'm not sure how to validate it works as expected other than allowing
the e2e tests to run, so if there is anything I can do, let me know :)