util/watch: Override GVK on incoming objects (#1163)
The K8s API server and client-go together have the behavior that objects
returned from List() calls do not have TypeMeta set.

For one-off List() requests this is fine becuase you already know the
type! But this interacts poorly with the generated implementations of
objects' .GetObjectKind().GroupVersionKind(), as those just directly
read from the TypeMeta fields (which again: are not set).

So this commit works around this behavior by getting the GVK at the
start of the Watch() call and explicitly setting it on all incoming
objects.