util/watch: Add (*Store[T]).Listen() method (#1163)
The Listen() method returns a util.BroadcastReceiver that will be
updated whenever the object is modified or deleted.

This is required *for now* for the stateless scheduler work, so that we
can be separately notified when there's changes to an object without
hooking deeper into the internal state.

We can probably remove this once the scheduler plugin's agnent request
handler server is removed, at the end of the stateless scheduler work.