ci: build neonvm-daemon in build-images.yaml (#1168)
Build and publish neonvm-daemon images since they are expected to be
used by neondatabase/neon repository during the "build vm image" step. [1]

For the simplicity, the neondatabase/autoscaling repo keeps building
local version of the daemon image during the vm-example workflow
(cronbased job) and during the build-test-vm workflow (is used as an
external workflow in the e2e tests).

Closes #1167

[1]: https://github.com/neondatabase/neon/blob/14c4fae64af5613c682ec7dd7d30e484c476e5af/.github/workflows/build_and_test.yml#L820

---------

Signed-off-by: Misha Sakhnov <misha@neon.tech>